pyproject.toml
[project]
name = "uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "boto3>=1.38.17",
    "chardet>=5.2.0",
    "docx2txt>=0.9",
    "fastapi>=0.115.12",
    "google-genai>=1.15.0",
    "httpx>=0.28.1",
    "numpy>=2.2.5",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.6",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "python-doctr>=0.11.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "streamlit>=1.45.1",
    "tabulate>=0.9.0",
    "uvicorn>=0.34.2",
    "vertexai>=1.71.1",
]
